requirements.txt
Django
djangorestframework
djangorestframework-jwt
djangorestframework_simplejwt
django-cors-headers
mysqlclient
gunicorn